requirements.txt
fastapi
uvicorn
jinja2
PyYAML
pydantic
pytest
pytest-asyncio
flake8
isort
mypy
